requirements.txt
fastapi
uvicorn
gunicorn
psycopg2-binary
requests
python-dotenv
pandas
numpy
